requirements.txt
asgiref==3.5.2
astroid==2.11.5
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
dill==0.3.5.1
Django==4.0.4
django-model-utils==4.2.0
djangorestframework==3.13.1
flake8==4.0.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.9
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
wrapt==1.14.1
yapf==0.32.0
